+++ lib/tri/query/db_compat.py
@@ -1,11 +1,13 @@
-from collections import OrderedDict
-
 from tri.declarative import setdefaults
 
-_variable_factory_by_django_field_type = None
+from tri.query import register_variable_factory
 
 
-def setup():
+def setup_db_compat():
+    setup_db_compat_django()
+
+
+def setup_db_compat_django():
     from tri.query import Variable
     from django.db.models import IntegerField, FloatField, TextField, BooleanField, AutoField, CharField, CommaSeparatedIntegerField, DateField, DateTimeField, DecimalField, EmailField, URLField, TimeField, ForeignKey, ManyToOneRel, ManyToManyField, ManyToManyRel
 
@@ -17,30 +19,23 @@ def setup():
         return Variable.choice_queryset(**kwargs)
 
     # The order here is significant because of inheritance structure. More specific must be below less specific.
-    global _variable_factory_by_django_field_type
-    _variable_factory_by_django_field_type = OrderedDict([
-        (CharField, lambda model_field, **kwargs: Variable(**kwargs)),
-        (URLField, lambda model_field, **kwargs: Variable.url(**kwargs)),
-        (TimeField, lambda model_field, **kwargs: Variable.time(**kwargs)),
-        (EmailField, lambda model_field, **kwargs: Variable.email(**kwargs)),
-        (DecimalField, lambda model_field, **kwargs: Variable.decimal(**kwargs)),
-        (DateField, lambda model_field, **kwargs: Variable.date(**kwargs)),
-        (DateTimeField, lambda model_field, **kwargs: Variable.datetime(**kwargs)),
-        (CommaSeparatedIntegerField, lambda model_field, **kwargs: Variable.comma_separated(parent_field=Variable.integer(**kwargs))),
-        (BooleanField, lambda model_field, **kwargs: Variable.boolean(**kwargs)),
-        (TextField, lambda model_field, **kwargs: Variable.text(**kwargs)),
-        (FloatField, lambda model_field, **kwargs: Variable.float(**kwargs)),
-        (IntegerField, lambda model_field, **kwargs: Variable.integer(**kwargs)),
-        (AutoField, lambda model_field, **kwargs: Variable.integer(**setdefaults(kwargs, dict(show=False)))),
-        (ManyToOneRel, None),
-        (ManyToManyField, lambda model_field, **kwargs: Variable.multi_choice_queryset(**setdefaults(kwargs, dict(choices=model_field.rel.to._default_manager.all())))),
-        (ManyToManyRel, None),
-        # (ManyToManyRel_Related, None),
-        (ForeignKey, foreign_key_factory),
-    ])
+    register_variable_factory(CharField, Variable),
+    register_variable_factory(URLField, Variable.url),
+    register_variable_factory(TimeField, Variable.time),
+    register_variable_factory(EmailField, Variable.email),
+    register_variable_factory(DecimalField, Variable.decimal),
+    register_variable_factory(DateField, Variable.date),
+    register_variable_factory(DateTimeField, Variable.datetime),
+    register_variable_factory(BooleanField, Variable.boolean),
+    register_variable_factory(TextField, Variable.text),
+    register_variable_factory(FloatField, Variable.float),
+    register_variable_factory(IntegerField, Variable.integer),
 
+    register_variable_factory(CommaSeparatedIntegerField, lambda **kwargs: Variable.comma_separated(parent_field=Variable.integer(**kwargs))),
+    register_variable_factory(AutoField, lambda **kwargs: Variable.integer(**setdefaults(kwargs, dict(show=False)))),
+    register_variable_factory(ManyToManyField, lambda model_field, **kwargs: Variable.multi_choice_queryset(model_field=model_field, **setdefaults(kwargs, dict(choices=model_field.rel.to._default_manager.all())))),
 
-def register_field_factory(field_class, factory):
-    _variable_factory_by_django_field_type[field_class] = factory
-
-setup()
+    register_variable_factory(ManyToOneRel, None),
+    register_variable_factory(ManyToManyRel, None),
+    # (ManyToManyRel_Related, None),
+    register_variable_factory(ForeignKey, foreign_key_factory),